[SCRUM-60] refactor: Impl suffix 수정
