+++ src/main/java/com/projectlyrics/server/domain/user/repository/CommandUserRepository.java
@@ -1,0 +1,6 @@
+package com.projectlyrics.server.domain.user.repository;
+
+import com.projectlyrics.server.domain.user.repository.impl.JpaUserCommandRepository;
+
+public interface CommandUserRepository extends JpaUserCommandRepository {
+}
