+++ src/main/java/com/projectlyrics/server/domain/user/repository/QueryUserRepository.java
@@ -4,7 +4,7 @@ import com.projectlyrics.server.domain.user.entity.User;
 import com.projectlyrics.server.domain.auth.external.AuthProvider;
 import java.util.Optional;
 
-public interface UserQueryRepository {
+public interface QueryUserRepository {
 
   Optional<User> findBySocialIdAndAuthProviderAndNotDeleted(long socialId, AuthProvider authProvider);
 }
