--- src/main/java/com/projectlyrics/server/domain/user/repository/UserCommandQueryRepository.java
@@ -1,4 +1,0 @@
-package com.projectlyrics.server.domain.user.repository;
-
-public interface UserCommandQueryRepository extends UserCommandRepository, UserQueryRepository {
-}
