+++ src/main/java/com/projectlyrics/server/domain/user/repository/impl/JpaUserCommandRepository.java
@@ -1,9 +1,9 @@
-package com.projectlyrics.server.domain.user.repository;
+package com.projectlyrics.server.domain.user.repository.impl;
 
 import com.projectlyrics.server.domain.user.entity.User;
 import org.springframework.data.repository.Repository;
 
-public interface UserCommandRepository extends Repository<User, Long> {
+public interface JpaUserCommandRepository extends Repository<User, Long> {
 
   User save(User entity);
 }
