+++ src/main/java/com/projectlyrics/server/domain/user/repository/impl/QueryDslQueryUserRepository.java
@@ -1,8 +1,9 @@
-package com.projectlyrics.server.domain.user.repository;
+package com.projectlyrics.server.domain.user.repository.impl;
 
 import com.projectlyrics.server.domain.user.entity.QUser;
 import com.projectlyrics.server.domain.user.entity.User;
 import com.projectlyrics.server.domain.auth.external.AuthProvider;
+import com.projectlyrics.server.domain.user.repository.QueryUserRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
-public class UserQueryRepositoryImpl implements UserQueryRepository {
+public class QueryDslQueryUserRepository implements QueryUserRepository {
 
   private final JPAQueryFactory jpaQueryFactory;
